refactor(layout): adopt standard Go project layout (cmd/ + internal/)
Move all Go source from the flat root package into:
- cmd/talos-mcp/    entry point, version vars, server wiring
- internal/talos/   Client struct (embeds *talosclient.Client), WithNodes
- internal/tools/   Handlers struct, all MCP tool handler methods

Key improvements:
- Package-level encapsulation enforced by internal/ visibility
- Embed *talosclient.Client so handlers call h.Client.Version(ctx) directly
- Move marshalResource/marshalResourceDefinition to tools (handler concern)
- Fix checkPathAllowed path traversal: add filepath.Clean() to canonicalize
  ".." components before prefix matching (pre-existing security bug)
- Add -coverpkg=./... for accurate cross-package coverage reporting
- Anchor binary pattern in .gitignore to /talos-mcp (was matching cmd/talos-mcp/)
- Update build path in Makefile, .goreleaser.yaml, ci.yml, and CLAUDE.md
