+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
       - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v5
         with:
           go-version-file: go.mod
-      - run: go test -v -race -coverprofile=coverage.out ./...
+      - run: go test -v -race -coverprofile=coverage.out -coverpkg=./... ./...
       - run: go tool cover -func=coverage.out
       - uses: codecov/codecov-action@b9fd7d16f6d7d1b5d2bec1a2887e65ceed900238 # v4
         with:
@@ -53,4 +53,4 @@ jobs:
       - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v5
         with:
           go-version-file: go.mod
-      - run: go build -o talos-mcp .
+      - run: go build -o talos-mcp ./cmd/talos-mcp
